requirements.txt
dj-database-url
Django
gunicorn
pytz
sqlparse
whitenoise
psycopg2
factory_boy